Added README.md and LICENCE. Moved ExrImage struct to file_formats module.
