+++ core-engine/src/file_formats/mod.rs
@@ -1,0 +1,3 @@
+pub mod exr;
+
+pub use exr::ExrImage;
